LayoutTests/fast/canvas/webgl/null-object-behaviour-expected.txt

Tests calling WebGL APIs without providing the necessary objects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.compileShader() threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.linkProgram() threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.attachShader() threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.attachShader(program, undefined) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.attachShader(undefined, shader) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.detachShader(program, undefined) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.detachShader(undefined, shader) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.shaderSource() threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.shaderSource(undefined, 'foo') threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.bindAttribLocation(undefined, 0, 'foo') threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.bindBuffer(context.ARRAY_BUFFER, 0) is undefined.
PASS context.bindFramebuffer(context.FRAMEBUFFER, 0) is undefined.
PASS context.bindRenderbuffer(context.RENDERBUFFER, 0) is undefined.
PASS context.bindTexture(context.TEXTURE_2D, 0) is undefined.
PASS context.framebufferRenderbuffer(context.FRAMEBUFFER, context.DEPTH_ATTACHMENT, context.RENDERBUFFER, 0) is undefined.
PASS context.framebufferTexture2D(context.FRAMEBUFFER, context.COLOR_ATTACHMENT0, context.TEXTURE_2D, 0, 0) is undefined.
PASS context.getProgramParameter(undefined, 0) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.getProgramInfoLog(undefined, 0) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.getShaderParameter(undefined, 0) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.getShaderInfoLog(undefined, 0) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.getShaderSource(undefined) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.getUniform(undefined, 0) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS context.getUniformLocation(undefined, 'foo') threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS successfullyParsed is true

TEST COMPLETE
